Add configurable GLMIntegration and update modules
